Tor memory management
- Reverted recent Tor changes.
